+++ pubspec.yaml
@@ -29,24 +29,24 @@ dependencies:
   intl: ^0.17.0
   location: ^4.3.0
   printing: ^5.6.6
-  geocoding: ^2.0.1
+  geocoding: ^2.0.2
   # file_picker: ^2.1.6
 
-  firebase_core: ^1.10.6
-  cloud_firestore: ^3.1.5
+  firebase_core: ^1.11.0
+  cloud_firestore: ^3.1.6
   # firebase_storage: ^10.0.1
-  firebase_auth: ^3.3.4
-  firebase_analytics: ^9.0.4
+  firebase_auth: ^3.3.5
+  firebase_analytics: ^9.0.5
 
-  google_sign_in: ^5.2.1
+  google_sign_in: ^5.2.3
   flutter_signin_button: ^2.0.0
   sign_in_with_apple: ^3.3.0
-  flutter_facebook_auth: ^3.5.7
+  flutter_facebook_auth: ^4.0.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.4
-  firebase_app_check: ^0.0.6+3
+  firebase_app_check: ^0.0.6+4
 
 dev_dependencies:
   flutter_test:
